+++ app/src/main/java/org/simple/clinic/enterotp/EnterOtpUi.kt
@@ -10,4 +10,6 @@ interface EnterOtpUi {
   fun showOtpEntryMode(mode: OtpEntryMode)
   fun showFailedAttemptOtpError(attemptsRemaining: Int)
   fun showLimitReachedError(attemptsMade: Int)
+  fun hideResendSmsButton()
+  fun showResendSmsButton()
 }